requirements.txt
# MOP AI Platform — Python Dependencies
#
# Install:  pip install -r requirements.txt

# Anthropic Claude API
anthropic>=0.30.0

# Pydantic v2 — canonical model and validation
pydantic>=2.0.0

# YAML config parsing
PyYAML>=6.0

# PDF parsing (pdfplumber preferred; PyPDF2 is the fallback)
pdfplumber>=0.10.0
PyPDF2>=3.0.0  # fallback

# DOCX parsing
python-docx>=1.1.0
lxml>=5.0.0          # XML/tracked-changes stripping for DOCX

# FastAPI — REST API + UI server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0   # ASGI server (includes websockets + httptools)
python-multipart>=0.0.9     # multipart/form-data for file uploads

# Phase 2 — Execution Engine
netmiko>=4.3.0          # SSH driver for all vendors
paramiko>=3.4.0         # SSH transport, jump host proxy
cryptography>=42.0.0    # credential file encryption
jinja2>=3.1.0           # HTML report rendering

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Optional OCR (uncomment to enable scanned-PDF support):
# pdf2image>=1.16.0
# pytesseract>=0.3.10
# Pillow>=10.0.0

# Optional Phase 2 extras (uncomment to enable):
# hvac>=2.1.0           # HashiCorp Vault credential store
# aiosmtplib>=3.0.0     # async SMTP for email notifications
# pdpyras>=5.1.0        # PagerDuty SDK
